+++ src/Householder.Server/Queries/GetAllResidents/GetAllResidentsQueryHandler.cs
@@ -1,0 +1,14 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using Householder.Server.Models;
+
+namespace Householder.Server.Queries
+{
+    public class GetAllResidentsQueryHandler : IQueryHandler<GetAllResidentsQuery, IEnumerable<Resident>>
+    {
+        public Task<IEnumerable<Resident>> Handle(GetAllResidentsQuery query)
+        {
+            throw new System.NotImplementedException();
+        }
+    }
+}