+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### Version 0.3.1, 2020-08-04
+ - Added queries for Resident
+ - Added commands for Resident
+
 ### Version 0.3.0, 2020-08-04
  - Added base queries and commands
  - Added LightInject dependency